requirements.txt
# Embeddings y ML
FlagEmbedding==1.2.5
torch>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector Store
chromadb>=0.4.0
faiss-cpu>=1.7.4

# LLM Local
ollama>=0.1.0
transformers>=4.35.0

# Backend
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Frontend
streamlit>=1.28.0
gradio>=4.0.0

# Utils
python-dotenv>=1.0.0
pandas>=2.1.0
regex>=2023.10.0
tqdm>=4.66.0